requirements.txt
python-dotenv
kagglehub
pillow

numpy
pandas

opencv-contrib-python
pathlib
matplotlib

imutils


# pytest
# pytest-cov



# torch
# torchvision